+++ examples/next-wagmi/app/page.tsx
@@ -2,8 +2,9 @@
 
 import { useState, useEffect } from "react";
 import "./page.css";
-import { SophonWidget } from "@sophon-labs/react";
+import { SophonWidget } from "@sophon-labs/account-react";
 import ExampleMethods from "@/components/methods";
+import { useAccount } from "wagmi";
 const checkIsDarkSchemePreferred = () => {
   if (typeof window !== "undefined") {
     return window.matchMedia?.("(prefers-color-scheme:dark)")?.matches ?? false;
@@ -14,8 +15,12 @@ const checkIsDarkSchemePreferred = () => {
 export default function Main() {
   const [isDarkMode, setIsDarkMode] = useState(checkIsDarkSchemePreferred);
 
+  const { address } = useAccount();
+
   useEffect(() => {
-    const darkModeMediaQuery = window.matchMedia("(prefers-color-scheme: dark)");
+    const darkModeMediaQuery = window.matchMedia(
+      "(prefers-color-scheme: dark)",
+    );
     const handleChange = () => setIsDarkMode(checkIsDarkSchemePreferred());
 
     darkModeMediaQuery.addEventListener("change", handleChange);
@@ -27,6 +32,12 @@ export default function Main() {
       <div className="modal">
         <SophonWidget variant="modal" />
         <ExampleMethods isDarkMode={isDarkMode} />
+
+        {address ? (
+          <span>Wagmi Account: {address}</span>
+        ) : (
+          <span>No Wagmi Account</span>
+        )}
       </div>
       <div className="footer">
         <div className="footer-text">Made with ❤️ by Sophon Team</div>