+++ examples/next-wagmi/lib/wagmi.ts
@@ -1,17 +1,19 @@
+import { sophonTestnet } from "wagmi/chains";
 import { http, createConfig } from "wagmi";
-import { sophonTestnet } from "@sophon-labs/react";
 
-export const config = createConfig({
-  chains: [sophonTestnet],
-  multiInjectedProviderDiscovery: false,
-  ssr: true,
-  transports: {
-    [sophonTestnet.id]: http(),
-  },
-});
+export function getConfig() {
+  return createConfig({
+    chains: [sophonTestnet],
+    multiInjectedProviderDiscovery: false,
+    ssr: true,
+    transports: {
+      [sophonTestnet.id]: http(),
+    },
+  });
+}
 
 declare module "wagmi" {
   interface Register {
-    config: typeof config;
+    config: ReturnType<typeof getConfig>;
   }
 }