+++ examples/sophon-demo/examples/hooks/context/SDKHasLoadedUsage.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useSophonContext } from "@sophon-labs/react";
+import { useSophonContext } from "@sophon-labs/account-react";
 
 export default async function ExampleHookUsage() {
   const { sdkHasLoaded } = useSophonContext();