+++ examples/sophon-demo/examples/ui/embedded/EmbeddedBorderBackground.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { SophonEmbeddedWidget } from "@sophon-labs/react";
+import { SophonEmbeddedWidget } from "@sophon-labs/account-react";
 
 export default function ExampleComponent() {
   return <SophonEmbeddedWidget background="with-border" />;