+++ examples/sophon-demo/examples/ui/sophon-widget/WidgetWithDropdown.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { SophonWidget } from "@sophon-labs/react";
+import { SophonWidget } from "@sophon-labs/account-react";
 
 export default function ExampleComponent() {
   return <SophonWidget variant="dropdown" label="Login With Dropdown" />;