+++ packages/account-wagmi/src/index.ts
@@ -1,0 +1,1 @@
+export { DynamicWagmiConnector as SophonWagmiConnector } from "@dynamic-labs/wagmi-connector";