+++ packages/react/src/ethereum.ts
@@ -1,2 +1,0 @@
-export { isEthereumWallet } from "@dynamic-labs/ethereum";
-export { isZKsyncConnector } from "@dynamic-labs/ethereum-aa";