+++ packages/react/src/wagmi.tsx
@@ -1,9 +1,0 @@
-import { DynamicWagmiConnector } from "@dynamic-labs/wagmi-connector";
-
-export const SophonWagmiConnector = ({
-  children,
-}: {
-  children: React.ReactNode;
-}) => {
-  return <DynamicWagmiConnector>{children}</DynamicWagmiConnector>;
-};