+++ packages/account-core/CHANGELOG.md
@@ -1,4 +1,46 @@
-# @sophon-labs/wallet
+# @sophon-labs/account-core
+
+## 1.1.6
+
+### Patch Changes
+
+- Bump version
+
+## 1.1.5
+
+### Patch Changes
+
+- Fixing type export
+
+## 1.1.4
+
+### Patch Changes
+
+- Exposing more types
+
+## 1.1.3
+
+### Patch Changes
+
+- Version bump
+
+## 1.1.2
+
+### Patch Changes
+
+- Version bump
+
+## 1.1.1
+
+### Patch Changes
+
+- Bumping version
+
+## 1.1.0
+
+### Minor Changes
+
+- Renaming packages, separating concerns.
 
 ## 1.0.13